Fix clang-tidy warnings and improve CI reporting (#3)
* Fix clang-tidy warnings and improve CI reporting

Fixed several static analysis issues identified by clang-tidy to improve code quality and type safety. Additionally, enhanced the CI pipeline to provide better visibility into static analysis results.

- Fixed narrowing conversion warnings in `Reflector.cpp` and `Transformer.cpp` using `static_cast<int>`.
- Resolved signed/unsigned comparison warnings in `RotorBox.cpp` using `std::cmp_not_equal` (C++20).
- Improved const-correctness in `RotorBox.cpp` by using `const auto&` in range-based loops.
- Updated `.github/workflows/static-analysis.yml` to generate a Markdown summary of clang-tidy warnings (total count and count by type) directly in the GitHub Actions Job Summary.

* Fix clang-tidy warnings in Rotor and Transformer

Cleaned up several static analysis findings in the core engine

- Fixed narrowing conversion warnings in `Rotor.cpp` and `Transformer.cpp` using explicit `static_cast<int>`.
- Resolved `readability-qualified-auto` and `readability-identifier-length` findings in `Rotor.cpp` by refining iterator types and naming

+++ RotorBox/Transformer/Reflector/src/Reflector.cpp
@@ -16,7 +16,7 @@ Reflector::Reflector(const ReflectorConfig& config) {
     }
 
     for (size_t i = 0; i < config.wiring.size(); ++i) {
-        setTransformValue(0, i, config.wiring[i]);
+        setTransformValue(0, static_cast<int>(i), config.wiring[i]);
     }
 
     // Initialize reverse transformation vector to -1.
